requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
python-nmap==0.7.1
requests==2.31.0
reportlab==4.1.0
urllib3==2.2.0
google-generativeai==0.3.2
supabase==2.3.0
